mk/pkginstall: DATASET-780, permit existing users
There should be an option with packages to re-use an existing user or
group, even if a specific uid/gid is provided.  This commit introduces a
user variable to configure this: PKGINSTALL_IGNORE_UIDGID.

When PKGINSTALL_IGNORE_UIDGID is true then pkg_install will not try to
create the user/group as specified.  If the user/group does not exist
but the specified uid/gid is, then the uid/gid will be stripped so that
the system can allocate one at random.