Add new source package: lang/gnat_util
Several tools are built with GNAT sources and they have to be the
exact source that built the Ada compiler that builds the tool.  To
ensure this, a common trick to the create a library with those sources
and link the library instead of using the embedded source files which
never match the FSF compiler properly.  The first tool this new
gcc 4.9.0-based library will be used for is devel/gprbuild-aux