Import p5-MooX-HandlesVia-0.001005 as devel/p5-MooX-HandlesVia.
MooX::HandlesVia is an extension of Moo's 'handles' attribute
functionality. It provides a means of proxying functionality from
an external class to the given attribute. This is most commonly
used as a way to emulate 'Native Trait' behavior that has become
commonplace in Moose code, for which there was no Moo alternative.