Update phonon-backend-vls to 0.8.1
Code style improvements.
VLC 2.2 compatibility
Changing the volume through the PulseAudio stream itself is not propagated
to Phonon AudioOutputs, meaning the application can not reflect volume
changes done through other applications (e.g. a volume mixer).
Changing the muteness through the PulseAudio stream is not propagated
either.
The PulseAudio media.role property (as well as derived properties) is
always set to 'video' making all Phonon application use the video category
for output device selection (e.g. as seen in the Phonon KCM).
Revised the logic used to come up with an icon name for libvlc.
Setting the libvlc app id is now only dependent on
QCoreApplication::applicationName being set (previously it was also
dependent on QCoreApplication::applicationVersion). This change was made to
accomodate the fact that most applications neglect to set a version.
Prevent cross-thread pause attempts on unpausable media as libvlc
internally will call stop when a player can not be paused. This could cause
a deadlock when using the paused-play feature with internet streams.