Update misc/py-anita to 1.34.  Changes from 1.33:
Do not require releases to contain install floppies, since amd64 no
longer has them.

The default value of the --boot-from option is now "cdrom" if the
install floppies are missing.

When reporting a failed download of optional files, be clearer that
this is not an error.  Also, don't treat arbitrary exceptions during
downloads as indicating a missing file, only the expected IOError.

Increase the default disk size from 768M to 1G, as 768M is no
longer sufficient for amd64.

New command line option --structured-log-file.  This is like
--structured-log but takes the name of a log file as an
argument, and the structured logging is in addition to, not
instead of, the traditional logging to stdout.

Increase installation timeout from 1 hour to 3 hours so that
a system built with DEBUG and LOCKDEBUG can be installed under
qemu without timing out.

Unconditionally print a greeting message including the version number
of startup, so that logs from automated anita runs will always
indicate the version used.

In --structured-log mode, log the string matched by expect() instead
of the return value, as the latter is always zero and therefore
conveys no useful information.