This is successor of databases/dbh. See   http://mail-index.netbsd.org/pkgsrc-users/2015/01/15/msg020923.html
Import libdbh2-5.0.16 as databases/libdbh2.

Disk based hashes is a method to create multidimensional binary trees on disk.
This library permits the extension of database concept to a plethora of
electronic data, such as graphic information. With the multidimensional binary
tree it is possible to mathematically prove that access time to any
particular record is minimized (using the concept of critical points from
calculus), which provides the means to construct optimized databases for
particular applications.

(From: doc/html/DBH.html)
Disk Based Hashtables (DBH) 64 bit ? Library to create and manage hash
tables residing on disk. Associations are made between keys and values
so that for a given a key the value can be found and loaded into memory
quickly. Being disk based allows for large and persistent hashes. 64 bit
support allows for hashtables with sizes over 4 Gigabytes on 32 bit
systems. Quantified key generation allows for minimum access time on
balanced multidimensional trees.