Fix build failure on Darwin < 10
GEGL tries to use OpenCL in a really weird way. It has (and installs!)
a local copy of OpenCL headers and then tries to find implementation
by dlopen'ing "libOpenCL.so". The problem is that it fails to compile
on Darwin < 10 by trying to include non-existent native OpenCL headers
too, which is in fact only optional.