(pkgsrc)  - Add LICENSE= modified-bsd (upstream)  - Update 5.2 to 5.5 ----------------- WHATSNEW in 5.5 ===============
In 5.4 the new NAT friendly behaviour was achieved by sending packets
with an rport of 0 to the server. In 5.5 this is no longer needed since
the server looks at the version number; hence 5.5 packets still go out
with rport set, making them compatible with servers of versions 5.3 or lower.
This was a small change/correction and 5.5 is a minor update.

WHATSNEW in 5.4
===============

DHIS packet messages leaving a host will now leave from the DHIS
return port (58800 by default) instead of a randomly assigned port.

Messages sent by this client to a server are marked with a return port of 0 in
the payload of the DHIS packet which causes the server to reply to the UDP port
from which it receives a packet. In other words, the return UDP port is no
longer embedded in a DHIS packet but instead is taken by the server from the
UDP layer of the network packet.

This feature (dhis client) only works with a DHIS server that implements it;
hence, a dhisd server of version 5.4 or above is required for a 5.4 client to
operate.

This feature is particularly useful in systems behind NAT; whereas before a
NAT redirect rule to the return DHIS port (58800 by default) was required, with
DHIS 5.4 this is no longer required since NAT boxes should keep the UDP state
and redirect back to the right host.

A new protocol is also implemented with DHID 5.4; instead of waiting for
server checks with CHECK_REQ, the client refreshes itself to the server
by sending periodic ECHO_REQ packets to it (every refresh period). The
server keeps the host online and only brings it offline it it doesn't
see any ECHO_REQ packets for up to 3 times the refresh period.

WHATSNEW in 5.3
===============

The documentation was updated.

DHID can now be executed under any user (not only root) without
changes to the code.

The pid file default location was changed to /var/run/dhid.pid