Update typo3_62 to 6.2.13 (TYPO3 6.2.13).
The TYPO3 community announces the release of TYPO3 CMS version 6.2.13
LTS, which is now ready for you to download.

This version is a maintenance release and contains bug fixes as well as
various improvements for the day-to-day administration of a TYPO3
website (Extension Manager and management of reference index).


PHP 5.6 support
- ---------------

Although the TYPO3 CMS Team aims at eventually supporting PHP 5.6 with
TYPO3 6.2 LTS, we are aware of some in-depth issues. As such, we
highly recommend to keep PHP 5.3 - 5.5 when running TYPO3 6.2 LTS for
the time being. Hopefully this should be fixed with the next release.


Image handling
- --------------

The base data used for the checksum calculation of processed files
have been changed. This should be transparent for you unless you are
having a large installation. In such case, we enjoin you to read the
details about this release (link below) and to make use of the
dedicated upgrade wizard.