Fix build of packages depending on mariadb55-client
When a package depends on mysql-client, in includes mk/mysql-buildlink3.mk
and may get buildlinked with mariadb55-client if MYSQL_VERSION_DEFAULT is
set to MARIADB55.

mariadb55-client's buildlink3.mk defined BUILDLINK_*.mariadb-client
variables, wich caused build problems with packages expecting
BUILDLINK_*.mysql-client to be defined. For instance, php-pdo_mysql
expectex BUILDLINK_PREFIX.mysql-client to be set and fails to build
without it.

Fix the problem by changing BUILDLINK_*.mariadb-client to
BUILDLINK_*.mysql-client in mariadb55-client/buildlink3.mk. Remove a
package specific woakrround for this exact ptoblem in php-mysqli.

NB: commit log on pkgsrc/databases/mariadb55-client/buildlink3.mk 1.2-1.3
and pkgsrc/databases/php-mysqli/Makefile 1.12-1.13 was truncated. This
change just adds a comment to the first file and a null-commit to the
second file to have an oportunity to provide the full cvs log.