Import paratype-ttf-20140616 as fonts/paratype-ttf.
The font families PT Sans and PT Serif were released in 2009-2010
with an open user license. The main aim of the project is to allow
the peoples of Russia to read and write in their native languages.
The project is dedicated to the 300-year anniversary of the civil
type invented by Peter the Great in 1708-1710 and was realized with
financial support from the Federal Agency for Press and Mass
Communications.

The fonts, beside the standard Western, Central European and Cyrillic
code pages, also contain characters of all title languages of the
Russian Federation.