Import Catch-1.3.4 as devel/catch
Catch stands for C++ Automated Test Cases in Headers and is a multi-paradigm
automated test framework for C++ and Objective-C (and, maybe, C). It is
implemented entirely in a set of header files, but is packaged up as a single
header for extra convenience.

Originally packaged in pkgsrc-wip (wip/catch) by Kamil Rytarowski