Update Padrino to 0.13.1.
Padrino 0.13.1 - Router and Reloader Updates, Ruby Compatibility, and Bug Fixes

			Posted on January 17, 2016 by Nathan Esquenazi

Padrino 0.13.0 was shipped 3 months ago on October 2015 and laid important
groundwork towards our eventual 1.0 release but also introduced some new
issues. After a few months of effort, we are excited to announce the release
of Padrino 0.13.1! This version is filled with routing and reloader
optimizations, compatibility updates, and bug fixes. Full details for this
release are below.

Router and Reloader Updates

The biggest improvement in this release is a significant reduction in the
memory usage of Padrino apps by changing the configuration of the underlying
mustermann router. Full list of improvements to router and reloader include:

* FIX #1975 Improve routing memory usage and performance (@namusyaka)
* FIX #1982 Support nested query for expanding path (@namusyaka)
* FIX #1978 Enable reloading of custom dependencies (@markglenfletcher)

Ruby Compatibility

Padrino 0.13.1 has been fixed to be fully compatible with Ruby 2.3 thanks to
@tyabe:

* FIX #2000 Fix mutex handling for Ruby 2.3 (@tyabe)

Bug Fixes and Miscellaneous

There are also several bug fixes and other updates:

* FIX Remove use of ActiveSupport in tests (@ujifgc)
* FIX #1994 Missing new line in mocha generator (@peter50216)
* FIX #1995 Invalid german dates (@ujifgc)
* FIX #1998 Only output a warning message if the spec task is invoked
  (@postmodern)
* FIX #1882 test for selected values for select tag (@ujifgc)
* FIX rendering exception for custom mime types (@nesquena)
* FIX relax mail gem dependency (@ujifgc)
* FIX minor doc typos (@lokyoung, @markglenfletcher)