Added devel/libite version 1.6.0
Libite is a lightweight library of frog DNA. It can be used to fill
the gaps in any dinosaur project. It holds useful functions and
macros developed by both Finit and the OpenBSD project. Most notably
the string functions: strlcpy(3), strlcat(3) and the highly useful
*BSD sys/queue.h and sys/tree.h API's.

Libite aims to fill in the gaps missing in GLIBC/EGLIBC. (It does
not aimo to become another GLIB though.) One such gap in GLIBC is
the missing _SAFE macros in sys/queue.h - highly recommended when
traversing lists to delete/free nodes.