Import py-patsy-0.4.1 as math/py-patsy.
Packaged for wip by Kamel Ibn Aziz Derouiche and myself.

A Python package for describing statistical models and for
Patsy is a Python library for describing statistical models
(especially linear models, or models that have a linear component)
and building design matrices. Patsy brings the convenience of R
formulas to Python.