Add creduce-2.5.0:
C-Reduce is a tool that takes a large C, C++, or OpenCL file that has a
property of interest (such as triggering a compiler bug) and
automatically produces a much smaller C/C++ file that has the same
property. It is intended for use by people who discover and report bugs
in compilers and other tools that process source code.