math/blas: Restore build on DragonFly
Linking blas with the gold linker fails with this error:
fatal error: --sysroot=: must take a non-empty argument

DragonFly has been using the gold linker by default for a while now.
Since I don't have time to track down this linking problem, I'm going
to restore the build on DragonFly by setting an environment variable
that forces DF to use the classic GNU linker instead.

At this point in time, I don't know if BLAS is unique or if LAPACK
and others need the same workaround.  If the latter, this variable
may need to be moved to the common LAPACK makefile.