Add lang/gcc6-aux to pkgsrc
I had intended to migrate the ada framework to use gcc6-aux instead of
lang/gcc5-aux, but unfortunately there's been a regression on NetBSD.
It builds and tests perfectly on DragonFly and FreeBSD, but it only
builds on NetBSD.  All the ACAT tests that involve tasking fail, but
the cause is unclear.  All the modifications used on gcc5-aux for NetBSD
are present in gcc6-aux but pthread destruction results in a segfault.

Since then gcc5-aux support has been improved, so I'm putting this away
for a while.  NetBSD has been removed from the platform support list
for now, but I'll review any patches I get to fix NetBSD tasking.

This compiler is based on gcc 6.2 and is primarily used for the Ada
frontend.