Add p5-String-Expand
This module implements utility functions for expanding embedded
variables in a string. Variable references are embedded in strings in
a similar form to the Bourne shell, namely, in the form $NAME or
${NAME}. In the former case, the NAME must consist of a capital letter
or underscore, and may be followed by zero or more capital letters,
digits or underscores. In the latter case, the name can consist of any
characters, but will be terminated by the first close brace character
'}'.