Import py-zstandard-0.7.0 as archivers/py-zstandard.
This project provides Python bindings for interfacing with the
Zstandard compression library. A C extension and CFFI interface are
provided.

The primary goal of the project is to provide a rich interface to
the underlying C API through a Pythonic interface while not sacrificing
performance. This means exposing most of the features and flexibility
of the C API while not sacrificing usability or safety that Python
provides.