Update php-redis to 2.2.8.
phpredis 2.2.8

The main improvement in this version of phpredis is support for Redis
Cluster. This version of phpredis is intended for versions of php older
than 7.

In addition there have been many bug fixes and improvements to non cluster
related commands, which are listed below.

I've attempted to include everyone who contribued to the project in each fix
description and have included names or github user ids.

Thanks to everyone for submitting bug reports and pull requests. A special
thanks to Remi Collet for helping with any and all packaging related issues

\o/

--- Improvements ---

* Added randomization to our seed nodes to balance which instance is used
to map the keyspace (Vitaliy Stepanyuk) [32eb1c5f]
* Added support for IPv6 addresses

--- Fixes ---

* PHP liveness checking workaround (Shafreeck Sea) [c18d58b9]
* Various documentation and code formatting and style fixes (ares333,
  sanpili, Bryan Nelson, linfangrong, Romero Malaquias, Viktor Sz?pe)
* Fix scan reply processing to use long instead of int to avoid overflow
  (mixiaojiong).
* Fix potential segfault in Redis Cluster session storage (Sergei Lomakov)
  [cc15aae]
* Fixed memory leak in discard function [17b1f427]
* Sanity check for igbinary unserialization (Maurus Cuelenaere) [3266b222,
  5528297a]
* Fix segfault occuring from unclosed socket connection for Redis Cluster
  (CatKang) [04196aee]
* Case insensitive zRangeByScore options
* Fixed dreaded size_t vs long long compiler warning